Get file path to penguins.csv and penguins_raw.csv files

palmerpenguins comes bundled with two csv files in its inst/extdata directory. This function make them easy to access.

path_to_file(path = NULL)

Arguments

path
Name of file in quotes with extension; "penguins.csv" and "penguins_raw.csv" will work. If NULL, the example files will be listed.

Source

This function is adapted from readxl::readxl_example().

Details

The data in each file is the same as penguins and penguins_raw, but without the variable types.

Examples

path_to_file()

#> [1] "penguins_raw.csv" "penguins.csv"    

path_to_file("penguins.csv")

#> [1] "/Users/runner/runners/2.169.1/work/_temp/Library/palmerpenguins/extdata/penguins.csv"

path_to_file("penguins_raw.csv")

#> [1] "/Users/runner/runners/2.169.1/work/_temp/Library/palmerpenguins/extdata/penguins_raw.csv"

if (FALSE) {
read.csv(path_to_file("penguins_raw.csv"))
}